Added some math and cryptology Python schoolwork
The caesar cipher is an excercise in the basic knowledge of cryptology,
it teaches the basics of how crypting systems have worked in history.

The collatz conjencture is still an unsolved problem in mathematics, it
has many names but is most commonly known as this. This was a basic
algorithmic exercise.